Further cleanup
