+++ UtilityBeltTests/UtilityBeltDataTests/DataModels/Entities/User+CoreDataProperties.swift
@@ -3,7 +3,6 @@
 import Foundation
 import CoreData
 
-
 extension User {
 
     @nonobjc public class func fetchRequest() -> NSFetchRequest<User> {
